requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
opencv-python-headless==4.13.0.92

# Auth
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Env
python-dotenv==1.0.1
# Vignette PNG generation
Pillow==10.4.0
numpy==2.0.2
